Move kotlin.text API misplaced in kotlin/native/internal folder
